requirements.txt
flask_cors
openai
flask
python-dotenv
requests
python-dateutil
